requirements.txt
beautifulsoup4
requests
pdfplumber
python-dateutil
rapidfuzz